Update and rename index (4).html to level :2 task 1 tribute page.html
A tribute page is an overview of someone whom we admire in our life. In this article, we are creating a tribute webpage of the Late A.P.J. Abdul Kalam Sir using HTML and CSS. The page features a clean and visually appealing design, incorporating a header with a centered title, a main section with content including text and images, and a footer for additional information or credits.